feat(v0.30): Format prices with zsh.01 precision, expand container width to 315px, and soften inactive tab overlay to 0x38000000

+++ common/src/main/java/com/example/advancedeconomics/shop/ShopItem.java
@@ -1,7 +1,7 @@
 package com.example.advancedeconomics.shop;
 
 /**
- * Shop Item Entry holding item identifier, display name, and base price.
+ * Shop Item record holding ID, Display Name, and decimal Base Price ($0.01 minimum).
  */
-public record ShopItem(String id, String displayName, long basePrice) {
+public record ShopItem(String id, String displayName, double basePrice) {
 }
